requirements.txt
opencv-python
ultralytics
tkinter
Pillow
matplotlib
gtts
playsound
numpy
